Use double values for all recorded metrics
Will fix https://github.com/uhafner/quality-monitor/issues/109

+++ src/main/java/edu/hm/hafner/grading/MetricStatistics.java
@@ -73,9 +73,10 @@ public class MetricStatistics {
      *
      * @return the metric values
      */
-    public Map<String, Integer> asMap() {
+    public Map<String, Double> asMap() {
         return projectValues.entrySet().stream()
-                .collect(Collectors.toMap(Map.Entry::getKey, entry -> entry.getValue().asInteger()));
+                .collect(Collectors.toMap(Map.Entry::getKey,
+                        entry -> entry.getValue().asDouble()));
     }
 
     @Override
